# app/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EpiLog API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/books/{id}": {
      "get": {
        "summary": "Alle Bücher im Shelf des eingeloggten Nutzers abrufen",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Book-ID (wird in dieser Route aktuell nicht verwendet)"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste der Bücher mit Lesefortschritt"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      },
      "post": {
        "summary": "Buch zum Shelf des Nutzers hinzufügen",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Book-ID (wird in dieser Route aktuell nicht verwendet)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "isbn"
                ],
                "properties": {
                  "isbn": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Buch erfolgreich hinzugefügt"
          },
          "400": {
            "description": "ISBN fehlt oder ungültig"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "409": {
            "description": "Buch bereits im Shelf"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      },
      "delete": {
        "summary": "Buch aus dem Shelf des Nutzers entfernen",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Book-ID (wird in dieser Route aktuell nicht verwendet, stattdessen Query-Parameter id)"
          },
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Book-ID des zu löschenden Eintrags"
          }
        ],
        "responses": {
          "200": {
            "description": "Buch erfolgreich entfernt"
          },
          "400": {
            "description": "Book-ID fehlt"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "404": {
            "description": "Buch nicht gefunden"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      }
    },
...
    "/api/users/{id}/readingactivity": {
      "get": {
        "summary": "Leseaktivität abrufen",
        "tags": [
          "ReadingActivity"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          },
          {
            "in": "query",
            "name": "min",
            "schema": {
              "type": "string"
            },
            "description": "Startdatum (z.B. 2026-01-01)"
          },
          {
            "in": "query",
            "name": "max",
            "schema": {
              "type": "string"
            },
            "description": "Enddatum (z.B. 2026-06-29)"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste der Leseaktivitäten"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      },
      "post": {
        "summary": "Leseaktivität anlegen oder Seiten hinzufügen",
        "tags": [
          "ReadingActivity"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pagesRead"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "description": "Optional, default ist heute"
                  },
                  "pagesRead": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leseaktivität erfolgreich angelegt"
          },
          "400": {
            "description": "Ungültige Anfragedaten"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      },
      "delete": {
        "summary": "Leseaktivität löschen",
        "tags": [
          "ReadingActivity"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date"
                ],
                "properties": {
                  "date": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leseaktivität erfolgreich gelöscht"
          },
          "400": {
            "description": "Ungültige Anfragedaten"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "404": {
            "description": "Nicht gefunden"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      }
    },
...
  },
...
}